marcketplace_microservices/products_api/products/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
pillow==10.4.0
pip==23.0.1
psycopg2==2.9.9
setuptools==65.5.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.44.0
